old_django_test_tasks/old_django_test_tasks/requirements.txt
beautifulsoup4
certifi
chardet
dj-database-url
Django
django-debug-toolbar
django-filter
django-money
djangorestframework
dry-rest-permissions
envparse
idna
lxml
Markdown
oauthlib
Pillow
psycopg2
psycopg2-binary
py-moneyed
pytz
requests
six
sqlparse
Unidecode<0.05
urllib3
django-ckeditor
awesome-slugify

